dedup_new/requirements.txt
# Core dependencies for OpenRouter DeepSeek deduplication

# OpenAI-compatible API client (OpenRouter uses OpenAI API format)
openai>=1.12.0

# Environment variable management
python-dotenv>=1.0.0

# YAML configuration
pyyaml>=6.0.1

# Data validation and serialization
pydantic>=2.5.0

# String similarity for duplicate detection
rapidfuzz>=3.5.0

# Markdown parsing
markdown>=3.5.1
beautifulsoup4>=4.12.0

# Progress bars and pretty output
tqdm>=4.66.0
rich>=13.7.0

# Code parsing
pygments>=2.17.0

# Optional: For embeddings-based similarity (if we want to add it later)
# sentence-transformers>=2.2.2
# scikit-learn>=1.3.0
